Move profile about component
