+++ src/features/profiles/ProfileAbout.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Button, Grid, Header, Tab } from "semantic-ui-react";
-import { Profile } from "../../../app/models/profile";
-import ProfileEditForm from "./ProfileEditForm";
+import { Profile } from "../../app/models/profile";
+import ProfileEditForm from "./form/ProfileEditForm";
 
 interface Props {
   profile: Profile;
